requirements.txt
oauth2client~=4.1.3
httplib2~=0.20.4
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-api-core==2.5.0
google-api-python-client==2.28.0
requests==2.26.0
celery~=5.1.2
redis==4.0.2
python-dotenv~=0.19.1
parsel~=1.6.0
